INTERNAL: Use GetResult interface in asyncGet api.
